fix: add newline character to notebook output
By default notebooks are stored with a newline at the end of the file. Furthermore, not including a newline will cause the pre-commit hook 'end-of-file-fixer' to fail and add a newline. This triggers a never-ending loop of fails. 
This fix will ensure that any adjusted notebook does include a newline at the end of the file, preserving original behaviour and reducing interference with other pre-commit hooks.